+++ src/main/java/Bob_BE/domain/discount/dto/data/DiscountDataDto.java
@@ -7,13 +7,4 @@ import lombok.NoArgsConstructor;
 
 public class DiscountDataDto {
 
-    @Getter
-    @Builder
-    @NoArgsConstructor
-    @AllArgsConstructor
-    public static class CreateDiscountDataDto {
-
-        Long menuId;
-        Integer discountPrice;
-    }
 }
