+++ src/main/java/Bob_BE/domain/discount/dto/response/DiscountResponseDto.java
@@ -15,7 +15,7 @@ public class DiscountResponseDto {
     @AllArgsConstructor
     public static class CreateDiscountResponseDto {
 
-        Long discountId;
-        LocalDateTime createdAt;
+        private Long discountId;
+        private LocalDateTime createdAt;
     }
 }
