+++ src/main/java/Bob_BE/domain/discountMenu/converter/DiscountMenuConverter.java
@@ -1,11 +1,7 @@
-package Bob_BE.domain.DiscountMenu.converter;
+package Bob_BE.domain.discountMenu.converter;
 
-import Bob_BE.domain.DiscountMenu.entity.DiscountMenu;
-import Bob_BE.domain.discount.dto.data.DiscountDataDto;
-import Bob_BE.domain.discount.entity.Discount;
+import Bob_BE.domain.discountMenu.entity.DiscountMenu;
 import Bob_BE.domain.menu.entity.Menu;
-
-import java.util.List;
 
 public class DiscountMenuConverter {
