+++ src/main/java/Bob_BE/global/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package Bob_BE.global.swagger;
+package Bob_BE.global.config;
 
 import io.swagger.v3.oas.models.Components;
 import io.swagger.v3.oas.models.OpenAPI;
